fix basic header

+++ rtmp/src/main/java/com/pedro/rtmp/rtmp/chunk/ChunkStreamId.kt
@@ -3,7 +3,7 @@ package com.pedro.rtmp.rtmp.chunk
 /**
  * Created by pedro on 21/04/21.
  */
-enum class ChunkStreamId(val mark: Byte) {
+enum class ChunkStreamId(val mark: Int) {
   PROTOCOL_CONTROL(0x02),
   OVER_CONNECTION(0x03),
   OVER_CONNECTION2(0x04),
